Moved withLock to ChangeLock interface
